chore: rename landing2 to iteam-mvp
